# make ansible key generator a tiny bit faster